ims-notifier/backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
jinja2==3.1.4
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.3.0
psycopg[binary]==3.2.3
python-multipart==0.0.12
google-auth==2.35.0
